requirements.txt
# LOFT Chat Backend Dependencies
# Generated from successful pip install

# FastAPI and server
fastapi[all]==0.115.7
uvicorn[standard]==0.35.0

# PydanticAI and OpenAI
pydantic-ai-slim[openai]==0.0.49
openai>=1.67.0

# HTTP client
httpx==0.28.1

# Environment and utilities
python-dotenv==1.0.0
sqlite-utils==3.38

# Database (PostgreSQL)
asyncpg==0.29.0
requests==2.32.3

# Core dependencies (auto-installed)
pydantic>=2.11.7
typing-extensions>=4.14.1
starlette>=0.45.3
click>=8.1.8
